requirements.txt
# Core LangChain - 版本锁定避免依赖冲突
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.73,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.9,<1.0.0

# LangChain Providers
langchain-openai==0.3.3
langchain-deepseek-official>=0.1.0
langchain-huggingface>=0.1.0,<1.0.0
langchain-chroma>=0.2.0,<0.3.0
langchain-milvus>=0.1.0,<0.3.0
langchain-unstructured>=0.1.0,<0.2.0

# Vector Stores - 版本锁定避免 tokenizers 冲突
chromadb>=0.5.0,<0.6.0
milvus-lite>=2.4.0
faiss-cpu>=1.8.0

# Embeddings - 版本锁定避免依赖冲突
sentence-transformers>=3.0.0
huggingface-hub>=0.20.0,<1.0.0
transformers>=4.0.0,<5.0.0
tokenizers<=0.20.3

# PDF & OCR
pypdf>=4.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10

# Unstructured (for PDF, DOCX, etc.)
unstructured[pdf]>=0.16.0
unstructured-client>=0.29.0
unstructured-inference>=0.8.0
onnxruntime<=1.19.2,>=1.17.0

# Reranking
cohere>=5.0.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.8.0
rank-bm25>=0.2.0
numpy<2.0.0

# Optional: Ollama
langchain-ollama>=0.2.0,<0.4.0

# Optional: Evaluation
ragas>=0.2.0
